feat: sum two numbers. + split blog into two parts
To keep each blog contain information about single topic, this blog has
been split into two parts. The first part will explain how you can take
information from the Frontend and Parse it in the backend, while second
part will explain how we can take information from the backend and
render it in a meaningful format in the frontend.

Further, to clarify that information from the frontend can be
processed, a small snippet has been added whereby a user can input two
numbers into the Form and the server will generate a sum and return it
as part of the sentence.