fix(footer): Relocated footer call to blog-page.js
The footer that I originally had was messing the whole design up.
I had placed in the layout page, due to which it was showing up in the
wrong place.
Relocated the `Footer` call to blog-page
removed the icons
changed the tag from <h4> to <h1>
removed unused code from layout.js