feat(blog): Comment Integration - Meta Tags
Utterances will use og:title for creating the issue comment
Utterances uses meta tags for title and description to create issue
description.
Example: https://github.com/ayushxx7/ayush-mandowara-blog/issues/10#issue-860451959
Changed the theme for utterances to photon-dark
Utterances will attach label comment on the issue

Passed in 'tags' for meta keywords to seo function
Title, Author, Keywords are fixed.

Note: Not going to fix description at the moment. This will be fixed in
a different issue. This is low priority anyway, since there is no
traffic on the website and the issues are not visible in the comment
section.

Closes #9