feat(search): search is working but not fully
search is only locating articles from the current page, problaby because
of how the site is setup to be broken down into multiple pages.

It might be better to have a separate component altogether which goes
through all the articles.

note: setting up the filter was surprisingly simple
Thanks to this blog:
https://aboutmonica.com/blog/create-gatsby-blog-search-tutorial